refactor SecurityService

+++ api/src/main/java/com/epam/esm/controller/OrderController.java
@@ -7,7 +7,6 @@ import com.epam.esm.service.SecurityService;
 import com.epam.esm.service.hateoas.HateoasService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -35,8 +34,8 @@ public class OrderController {
      */
     @PostMapping
     @RolesAllowed({"ADMIN", "USER"})
-    public ResponseEntity<OrderDto> createOrder(@Valid @RequestBody OrderDto orderDto, Authentication authentication) {
-        securityService.ifUserIdNotMatchingThrowException(authentication, orderDto.getUser().getId());
+    public ResponseEntity<OrderDto> createOrder(@Valid @RequestBody OrderDto orderDto) {
+        securityService.validateUserAccess(orderDto.getUser().getId());
         OrderDto createdOrderDto = orderService.createOrder(orderDto);
         hateoasService.attachHateoas(createdOrderDto);
         return ResponseEntity.status(CREATED).body(createdOrderDto);
@@ -50,8 +49,8 @@ public class OrderController {
      */
     @PutMapping
     @RolesAllowed({"ADMIN", "USER"})
-    public ResponseEntity<OrderDto> updateOrder(@Valid @RequestBody OrderDto orderDto, Authentication authentication) {
-        securityService.ifUserIdNotMatchingThrowException(authentication, orderDto.getUser().getId());
+    public ResponseEntity<OrderDto> updateOrder(@Valid @RequestBody OrderDto orderDto) {
+        securityService.validateUserAccess(orderDto.getUser().getId());
         OrderDto updatedOrderDto = orderService.updateOrder(orderDto);
         hateoasService.attachHateoas(updatedOrderDto);
         return ResponseEntity.ok().body(updatedOrderDto);
@@ -65,10 +64,9 @@ public class OrderController {
      */
     @DeleteMapping("/{orderId}")
     @RolesAllowed({"ADMIN", "USER"})
-    public ResponseEntity<GiftCertificateDto> deleteOrder(@PathVariable("orderId") @Min(1) Long orderId,
-                                                          Authentication authentication) {
+    public ResponseEntity<GiftCertificateDto> deleteOrder(@PathVariable("orderId") @Min(1) Long orderId) {
         OrderDto orderDto = orderService.getOrderById(orderId);
-        securityService.ifUserIdNotMatchingThrowException(authentication, orderDto.getUser().getId());
+        securityService.validateUserAccess(orderDto.getUser().getId());
         orderService.deleteOrder(orderId);
         return ResponseEntity.noContent().build();
     }
